Fix link to PNG home, remove reference to fribidi.
Tue Dec  5 15:49:31 2000  Owen Taylor  <otaylor@redhat.com>

	* INSTALL: Fix link to PNG home, remove reference
	to fribidi.

	* gdk/x11/gdkevents-x11.c: Fix up for new GMain API.

	* gdk/win32/gdkevents-win32.c: Fix up for new GMain API. (Not
	tested.)

	* gtk/gtkcheckmenuitem.c: Permanently turn on
	always_show_toggle.

	* gtk/gtkclist.c (gtk_clist_get_pixtext): Allow mask to
	be NULL. (Vladimir Klebanov, #26545)

	* gtk/gtkspinbutton.c (gtk_spin_button_insert_text): Fix
	some signed/unsigned comparison problems. (#6510,
	David Kaelbling)

	* gtk/gtkwidget.[ch] (gtk_widget_is_ancestor): Change
	return type to boolean (Oskar Liljeblad, #18648)

+++ gtk/gtkspinbutton.c
@@ -1209,7 +1209,7 @@ gtk_spin_button_insert_text (GtkEntry    *entry,
       gint i;
       GdkWChar pos_sign;
       GdkWChar neg_sign;
-      guint entry_length;
+      gint entry_length;
 
       entry_length = entry->text_length;
 
@@ -1244,7 +1244,7 @@ gtk_spin_button_insert_text (GtkEntry    *entry,
 	  }
 
       if (dotpos > -1 && *position > dotpos &&
-	  spin->digits - entry_length
+	  (gint)spin->digits - entry_length
 	    + dotpos - new_text_length + 1 < 0)
 	return;
 
@@ -1260,7 +1260,7 @@ gtk_spin_button_insert_text (GtkEntry    *entry,
 	    {
 	      if (!spin->digits || dotpos > -1 || 
  		  (new_text_length - 1 - i + entry_length
-		    - *position > spin->digits)) 
+		    - *position > (gint)spin->digits)) 
 		return;
 	      dotpos = *position + i;
 	    }
